Separated Library class from main.cpp
