+++ LibraryManager/src/Classes/BookClass.h
@@ -27,5 +27,4 @@ private:
 	size_t id;
 	short rating;
 
-
 };
